requirements.txt
matplotlib
matplotlib-inline
numpy
Pillow==9.0.1
pytorch-lightning==1.7.1
PyYAML==6.0
scikit-image
scikit-learn
scipy
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.7.2
torchvision==0.11.3+cu113
tqdm==4.64.0